mem_free 4GB wraparound on MacOSX

+++ source/libs/uti/sge_loadmem.c
@@ -817,7 +817,7 @@ int sge_loadmem(sge_mem_info_t *mem_info)
     mem_info->mem_total = mem_total / (1024*1024);
 
     host_statistics(mach_host_self (), HOST_VM_INFO, (host_info_t)&vm_info, &info_count);
-    mem_info->mem_free = vm_info.free_count*vm_page_size / (1024*1024);
+    mem_info->mem_free = ((double)vm_info.free_count)*vm_page_size / (1024*1024);
 
     return 0;
 }
